fix: Android NDK compatibility for libusb bindings
- Replace C.ulong with C.size_t for malloc calls to match Android NDK types
- Fix flexible array member access for iso_packet_desc using pointer arithmetic
- Include whitespace formatting fixes

These changes ensure compatibility with Android's NDK headers where:
- malloc expects size_t instead of ulong
- iso_packet_desc is defined as iso_packet_desc[LIBUSB_FLEXIBLE_ARRAY]

The fixes are backward compatible with other platforms.

+++ pkg/transfers/streaming_interface.go
@@ -96,7 +96,7 @@ func (si *StreamingInterface) ClaimFrameReader(formatIndex, frameIndex uint8) (*
 	vpcc := &descriptors.VideoProbeCommitControl{}
 	size := 48
 
-	buf := C.malloc(C.ulong(size))
+	buf := C.malloc(C.size_t(size))
 	defer C.free(buf)
 
 	// get the bounds
